# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.OrderRequest:
    properties:
      player_id:
        example: "123456789"
        type: string
      product_id:
        example: 1
        type: integer
      server_id:
        example: "1234"
        type: string
    type: object
  controllers.OrderStatusRequest:
    properties:
      status:
        example: PROCESSING
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data:
        description: Data yang dikirim API.
      message:
        description: Message dari API.
        type: string
      status:
        description: |-
          Status response.
          Contoh: success / error
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: REST API untuk mengelola game, product, provider, dan order top-up.
  title: Topup Games API
  version: "1.0"
paths:
...
  /api/orders:
    get:
      description: Mengambil daftar seluruh order.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil semua order
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Membuat order top-up baru berdasarkan product dan ID player.
      parameters:
      - description: Data order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Membuat order
      tags:
      - orders
  /api/orders/{id}:
    get:
      description: Mengambil satu order berdasarkan ID.
      parameters:
      - description: ID order
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil order berdasarkan ID
      tags:
      - orders
  /api/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Mengubah status order berdasarkan ID.
      parameters:
      - description: ID order
        in: path
        name: id
        required: true
        type: integer
      - description: Status order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengubah status order
      tags:
      - orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Masukkan token dengan format "Bearer {token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"